Green with Tests

+++ src/main/java/com/example/u1m6summative/model/InvoiceItem.java
@@ -1,6 +1,5 @@
 package com.example.u1m6summative.model;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 
 public class InvoiceItem {
@@ -9,8 +8,8 @@ public class InvoiceItem {
     private int invoiceId;
     private int itemId;
     private int quantity;
-    private BigDecimal unitRate;
-    private BigDecimal discount;
+    private Double unitRate;
+    private Double discount;
 
     public int getInvoiceItemId() {
         return invoiceItemId;
@@ -44,19 +43,19 @@ public class InvoiceItem {
         this.quantity = quantity;
     }
 
-    public BigDecimal getUnitRate() {
+    public Double getUnitRate() {
         return unitRate;
     }
 
-    public void setUnitRate(BigDecimal unitRate) {
+    public void setUnitRate(Double unitRate) {
         this.unitRate = unitRate;
     }
 
-    public BigDecimal getDiscount() {
+    public Double getDiscount() {
         return discount;
     }
 
-    public void setDiscount(BigDecimal discount) {
+    public void setDiscount(Double discount) {
         this.discount = discount;
     }
